create separate files for Generic* classes
